+++ packages/catlog/src/simulate/ode/polynomial.rs
@@ -103,15 +103,9 @@ where
     {
         self.components
             .iter()
-            .map(|(var, poly)| {
-                let var = var.to_string();
-                let var = if var.chars().count() > 1 {
-                    format!("\\text{{{var}}}")
-                } else {
-                    var
-                };
-                let lhs = format!("\\frac{{\\mathrm{{d}}}}{{\\mathrm{{d}}t}} {var}");
-                LatexEquation { lhs, rhs: poly.to_latex() }
+            .map(|(var, poly)| LatexEquation {
+                lhs: format!("\\frac{{\\mathrm{{d}}}}{{\\mathrm{{d}}t}} {var}"),
+                rhs: poly.to_latex(),
             })
             .collect()
     }